requirements.txt
django==3.2
djangorestframework
mysqlclient
cloudinary
python-decouple
djoser==2.1.0
django-admin-tailwind
django-cors-headers==3.7.0
django-environ==0.4.5
social-auth-app-django==4.0.0
django-paypal
pytz==2021.3

Pillow==8.1.2
mutagen==1.45.1 

psycopg2==2.9.2 

argon2-cffi==21.1.0 


stripe==2.58.0
braintree==4.13.1

pyJWT==2.1.0

boto3==1.18.63
django-storages==1.12.2 

markdown==3.3.4
gunicorn==20.1.0
whitenoise==5.3.0

requests==2.26.0